moved build macros to utils
