+++ zingolib/src/utils.rs
@@ -2,3 +2,37 @@
 
 pub mod conversion;
 pub mod error;
+
+macro_rules! build_method {
+    ($name:ident, $localtype:ty) => {
+        #[doc = "Set the $name field of the builder."]
+        pub fn $name(&mut self, $name: $localtype) -> &mut Self {
+            self.$name = Some($name);
+            self
+        }
+    };
+}
+#[cfg(test)] // temporary test gate as no production builders use this macros yet
+macro_rules! build_method_push {
+    ($name:ident, $localtype:ty) => {
+        #[doc = "Push a $ty to the builder."]
+        pub fn $name(&mut self, $name: $localtype) -> &mut Self {
+            self.$name.push($name);
+            self
+        }
+    };
+}
+#[cfg(test)] // temporary test gate as no production builders use this macros yet
+macro_rules! build_push_list {
+    ($name:ident, $builder:ident, $struct:ident) => {
+        for i in &$builder.$name {
+            $struct.$name.push(i.build());
+        }
+    };
+}
+
+pub(crate) use build_method;
+#[cfg(test)]
+pub(crate) use build_method_push;
+#[cfg(test)]
+pub(crate) use build_push_list;
