+++ zingolib/src/wallet/notes/sapling.rs
@@ -232,7 +232,8 @@ pub mod mocks {
     use zcash_primitives::{memo::Memo, transaction::TxId};
 
     use crate::{
-        mocks::{build_method, SaplingCryptoNoteBuilder},
+        mocks::SaplingCryptoNoteBuilder,
+        utils::build_method,
         wallet::{notes::ShieldedNoteInterface, traits::FromBytes},
     };
