+++ zingolib/src/wallet/transaction_record.rs
@@ -553,10 +553,10 @@ pub mod mocks {
 
     use crate::{
         mocks::{
-            build_method, build_method_push, build_push_list,
             nullifier::{OrchardNullifierBuilder, SaplingNullifierBuilder},
             random_txid,
         },
+        utils::{build_method, build_method_push, build_push_list},
         wallet::{
             data::mocks::OutgoingTxDataBuilder,
             notes::{
